requirements.txt
aiohttp==3.8.1
aiohttp-retry==2.4.6
aiosignal==1.2.0
altair==4.2.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.1
astor==0.8.1
asttokens==2.0.5
async-timeout==4.0.2
asyncssh==2.8.1
atpublic==3.0.1
attrs==21.4.0
backcall==0.2.0
base58==2.1.1
binaryornot==0.4.4
black==21.12b0
bleach==4.1.0
blinker==1.4
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
cookiecutter==1.7.3
cryptography==36.0.1
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.4.0
distro==1.6.0
dpath==2.0.5
dulwich==0.20.32
dvc==2.9.3
entrypoints==0.3
executing==0.8.2
flatten-dict==0.4.2
flufl.lock==7.0
frozenlist==1.3.0
fsspec==2022.1.0
ftfy==6.0.3
funcy==1.17
future==0.18.2
gitdb==4.0.9
GitPython==3.1.26
grandalf==0.6
idna==3.3
imbalanced-learn==0.9.0
imblearn==0.0
ipykernel==6.7.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jinja2-time==0.2.0
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
mailchecker==4.1.10
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==6.0.2
mypy-extensions==0.4.3
nanotime==0.5.2
nbclient==0.5.10
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
notebook==6.4.8
numpy==1.22.1
packaging==21.3
pandas==1.4.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
phonenumbers==8.12.41
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
ply==3.11
poyo==0.5.0
prometheus-client==0.13.0
prompt-toolkit==3.0.24
protobuf==3.19.3
psutil==5.9.0
pure-eval==0.2.2
pyarrow==6.0.1
pyasn1==0.4.8
pycparser==2.21
pydeck==0.7.1
pydot==1.4.2
pygit2==1.7.2
Pygments==2.11.2
pygtrie==2.4.2
Pympler==1.0.1
pyparsing==3.0.7
pyrsistent==0.18.1
python-benedict==0.24.3
python-dateutil==2.8.2
python-fsutil==0.6.0
python-slugify==5.0.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
rich==11.0.0
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.2
scipy==1.7.3
scmrepo==0.0.7
Send2Trash==1.8.0
shortuuid==1.0.8
shtab==1.5.3
six==1.16.0
smmap==5.0.0
stack-data==0.1.4
streamlit==1.4.0
tabulate==0.8.9
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.3
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
validators==0.18.2
voluptuous==0.12.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
xmltodict==0.12.0
yarl==1.7.2
zc.lockfile==2.0
